pyproject.toml
[project]
name = "prk-mcp-storyforge-agent-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
]
